AssetHeader base class for all asset types
- Magic number, versioning, and size are now incorporated into
  AssetHeader base class that every asset type derives.
- Magic numbers are now 4cc's
- Endianess check for 4cc and can be used in other places we find use
- Fixed issue where a version change in a dependent type wasn't
  forcing a recook of dependent files.  For example, if we rev'd the
  Image cooker version, all .tgas are recooked, but .spr and .atl
  files that depend on .tgas weren't recooked.

+++ src/chef/Cooker.h
@@ -41,7 +41,7 @@ class Cooker
 public:
     virtual ~Cooker() {}
 
-    u32 version() const { return mVersion; }
+    u16 version() const { return mVersion; }
 
     typedef List<kMEM_Chef, ChefString> ExtList;
     const ExtList & rawExts() const { ASSERT(mCookedExts.size() > 0); return mRawExts; }
@@ -51,7 +51,7 @@ public:
 
 protected:
     // Increase version in subclasses to force a recook of those asset types.
-    u32 mVersion = 0;
+    u16 mVersion = 0;
 
     ExtList mRawExts;
     ExtList mCookedExts;
